+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Most of the flake8 errors are resolved.
 - File checks and exception catching when loading points from a file.
 - Support for ',' delimiter for loading the data.
+- Support for loading the data from a 'ng_trajectory' csv file.
 
 ### Changed
 - Node is now executable using Python 3 as well.